Only create test user once

+++ server/test/lib.js
@@ -67,39 +67,45 @@ function statusAndJSON(verb, url, cookieFn, body, expectedStatus, after) {
 	};
 }
 
+var hasInitialized = false;
+
 module.exports = {
 	getCookie: getCookie,
 	statusAndJSON: statusAndJSON,
 	init: function() {
-		before(function(done) {
-			_users.withoutPermission.db = new dbUsers({ name: "Test User 1", email: "em@il.com", local: { username:_users.withoutPermission.username, secret: bcrypt.hashSync(_users.withoutPermission.password) }, permissions: { "links": false, "officers": false, "shows": false, "classes": false, "pictures": false, "calendar": false, "users": false }});
-			_users.withPermission.db = new dbUsers({ name: "Test User 2", email: "em@il.com", local: { username: _users.withPermission.username, secret: bcrypt.hashSync(_users.withPermission.password) }, permissions: { "links": true, "officers": true, "shows": true, "classes": true, "pictures": true, "calendar": true, "users": true }});
+		if(!hasInitialized) {
+			hasInitialized = true;
 
-			_users.withoutPermission.db.save(function() {
-				_users.withPermission.db.save(done);
-			});
-		});
+			before(function(done) {
+				_users.withoutPermission.db = new dbUsers({ name: "Test User 1", email: "em@il.com", local: { username:_users.withoutPermission.username, secret: bcrypt.hashSync(_users.withoutPermission.password) }, permissions: { "links": false, "officers": false, "shows": false, "classes": false, "pictures": false, "calendar": false, "users": false }});
+				_users.withPermission.db = new dbUsers({ name: "Test User 2", email: "em@il.com", local: { username: _users.withPermission.username, secret: bcrypt.hashSync(_users.withPermission.password) }, permissions: { "links": true, "officers": true, "shows": true, "classes": true, "pictures": true, "calendar": true, "users": true }});
 
-		before(function(done) {
-			request.post("http://127.0.0.1:9931/auth/local", { form: { username: _users.withoutPermission.username, password: _users.withoutPermission.password }}, function(err, response, body) {
-				var cookie = response.headers["set-cookie"].toString();
-				_users.withoutPermission.cookie = cookie.substr(0, cookie.indexOf(";"));
-				done();
+				_users.withoutPermission.db.save(function() {
+					_users.withPermission.db.save(done);
+				});
 			});
-		});
 
-		before(function(done) {
-			request.post("http://127.0.0.1:9931/auth/local", { form: { username: _users.withPermission.username, password: _users.withPermission.password }}, function(err, response, body) {
-				var cookie = response.headers["set-cookie"].toString();
-				_users.withPermission.cookie = cookie.substr(0, cookie.indexOf(";"));
-				done();
+			before(function(done) {
+				request.post("http://127.0.0.1:9931/auth/local", { form: { username: _users.withoutPermission.username, password: _users.withoutPermission.password }}, function(err, response, body) {
+					var cookie = response.headers["set-cookie"].toString();
+					_users.withoutPermission.cookie = cookie.substr(0, cookie.indexOf(";"));
+					done();
+				});
 			});
-		});
 
-		after(function(done) {
-			_users.withoutPermission.db.remove(function() {
-				_users.withPermission.db.remove(done);
+			before(function(done) {
+				request.post("http://127.0.0.1:9931/auth/local", { form: { username: _users.withPermission.username, password: _users.withPermission.password }}, function(err, response, body) {
+					var cookie = response.headers["set-cookie"].toString();
+					_users.withPermission.cookie = cookie.substr(0, cookie.indexOf(";"));
+					done();
+				});
 			});
-		});
+
+			after(function(done) {
+				_users.withoutPermission.db.remove(function() {
+					_users.withPermission.db.remove(done);
+				});
+			});
+		}
 	}
 };
